Switch to swizzling IDEMediaLibraryCompletionStrategy (#4)
Swizzling `IDEMediaResourceCompletionItem` and replacing the `name`
selector to return nil _almost_ worked in all cases. Unfortunately in
one codepath the `name` returned from this function was used to
initialize a `NSAttributedString`, resulting in a crash.

This replacement approach goes a level up and creates an initializer for
the `IDEMediaLibraryCompletionStrategy`, which vends the completion items,
and returns nil.

If this failed in theory you could just edit the `IDEKit.xcplugindata`
and delete the `IDEMediaLibraryCompletionStrategy` entry from there,
since that's how Xcode manages plugins, but this is much less
destructive so as long as it works!